src/requirements.txt
absl-py==1.4.0
accelerate==0.16.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
blobfile==2.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cmake==3.25.0
datasets==2.10.0
dill==0.3.6
filelock==3.9.0
frozenlist==1.3.3
fsspec==2023.1.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
grpcio==1.51.3
huggingface-hub==0.12.1
idna==3.4
Jinja2==3.1.2
langchain==0.0.115
lit==15.0.7
loralib==0.1.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
networkx==3.0rc1
numpy==1.24.1
oauthlib==3.2.2
openai==0.27.2
packaging==23.0
pandas==1.5.3
Pillow==9.3.0
protobuf==4.22.0
psutil==5.9.4
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodomex==3.17
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
six==1.16.0
sympy==1.11.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tiktoken==0.2.0
tokenizers==0.13.2
torchinfo==1.7.2
# Download your own torch nightly
# torch==2.0.0.dev20230228+cu117
# torchaudio==2.0.0.dev20230301+cu117
# torchvision==0.15.0.dev20230227+cu117
sentencepiece==0.1.99
tqdm==4.64.1
transformers==4.26.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.3
xxhash==3.2.0
yapf==0.32.0
yarl==1.8.2
wandb==0.16.1
